pyproject.toml
[tool.poetry]
name = "swertres"
version = "0.1.0"
description = ""
authors = ["Zykie Rallos <zykie1018@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
django-htmx = "^1.17.2"
django-extensions = "^3.2.3"
psycopg2 = "^2.9.9"
Pillow = "^9.3.0"
psycopg2-binary = "^2.9.5"
django-environ = "^0.11.2"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
coverage = {version = "^6.5", extras = ["toml"]}
django-coverage-plugin = "^2.0.4"
factory-boy = "^3.2.1"
ipdb = "^0.13.9"
pre-commit = "^2.20.0"
pytest = "^7.2.0"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pytest
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "df2_wiki.settings.base"

# isort
[tool.isort]
line_length = 88
profile = "black"
known_django = "django"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
multi_line_output = 3
skip_glob = "**/migrations/*.py"
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true


# coverage
[tool.coverage.run]
omit = [
    "*migrations*",
    "*tests*",
    "*staticfiles*",
    "*__init__.py",
    "df2_wiki/*",
    "conftest.py",
    "manage.py",
]
plugins = ["django_coverage_plugin"]

[tool.coverage.report]
omit = [
    "*migrations*",
    "*staticfiles*",
    "*tests*",
    "*__init__.py",
    "df2_wiki/*",
    "conftest.py",
    "manage.py",
]
